chore(deps): remove puppeteer (#29079)
Issue number: N/A
---------

<!-- Please do not submit updates to dependencies unless it fixes an
issue. -->

<!-- Please try to limit your pull request to one type (bugfix, feature,
etc). Submit multiple pull requests if needed. -->

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->

we have puppeteer as a dep of the project. this isn't a problem per se,
as it isn't actually used. however, we're hitting intermittent errors in
CI when puppeteer tries to pull down the latest version of chrome.

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

remove puppeteer as a dependency for the project. this is no longer
needed, as https://github.com/ionic-team/stencil/issues/4526 has been
fixed as of stencil v4.8.2

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

<!--
  If this introduces a breaking change:
1. Describe the impact and migration path for existing applications
below.
  2. Update the BREAKING.md file with the breaking change.
3. Add "BREAKING CHANGE: [...]" to the commit description when merging.
See
https://github.com/ionic-team/ionic-framework/blob/main/.github/CONTRIBUTING.md#footer
for more information.
-->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->