feat(textarea): add --highlight-height variable (#29063)
Issue number: internal

---------

## What is the current behavior?
In v7, using the legacy syntax, the height of the highlight on an item
could be adjusted using the `--highlight-height` variable. This variable
was not added to textarea and therefore would not work using the modern
syntax.

## What is the new behavior?
Adds the `--highlight-height` variable so that developers can customize
this height in `md` mode. Since the highlight element is not added for
`ios` mode, this variable won't do anything for `ios`. Note that this
diverges from the v7 behavior, where setting `--highlight-height`
enabled the highlight for `ios`. A design document outlining this has
been proposed here:
https://github.com/ionic-team/ionic-framework-design-documents/pull/252

## Does this introduce a breaking change?

- [ ] Yes
- [x] No

---------

Co-authored-by: ionitron <hi@ionicframework.com>